% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dirtylapply}
\alias{dirtylapply}
\title{dirtylapply replicates R's lapply: it takes an R list and an R function,
applies the function to each element, and returns a new list of results.}
\usage{
dirtylapply(list, func)
}
\description{
dirtylapply replicates R's lapply: it takes an R list and an R function,
applies the function to each element, and returns a new list of results.
}
